This is not how any of this works. Fix version detection
Don't support building DLLs OOTB so we don't have to deal with MSVC's
shittiness

Bloaty OS detection is utterly useless, remove it

Use ANSI versions of WinAPI functions

Fix Windows version detection. Check for kernel32.dll's platform version
because GetVersion is deprecated on Win8+

`system::kernel_t::name` and `system::kernel_t::friendly_name` are
literally useless as they're concatenations of already parsed and
returned data

Don't autodetect graphics frameworks, the defines are provided
externally as features *not* always autodetected

TODO "configMakefile": Ask @ThePhD if he's using D3D11 or D3D12

+++ src/cpu/frequency/windows.cpp
@@ -21,7 +21,7 @@
 unsigned int iware::cpu::frequency() noexcept {
 	LARGE_INTEGER freq;
 	QueryPerformanceFrequency(&freq);
-	return static_cast<unsigned int>(freq.QuadPart * 1000);
+	return freq.QuadPart * 1000;
 }
